+++ CHANGELOG.md
@@ -48,6 +48,7 @@
   - The buttons on the character sheet header now open the Ancestry, Background, and Class compendiums.
 - Items with advancements can no longer be mass-created as part of dropping a folder onto an actor sheet. (#736)
 - Adjusted the display of Power Roll Effects, including adding an image property.
+- Expired Active Effects are now labeled "Expired" in their duration rather than "None".
 
 ### Fixed